pyproject.toml
[project]
name = "vocalalchemy-backend"
version = "0.1.0"
requires-python = ">=3.11,<4.0"
dependencies = [
  "fastapi",
  "gunicorn",
  "uvicorn[standard]",
  "requests",
  "python-dotenv",
  "python-multipart",
  "jinja2",
  "sqlalchemy",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==3.2.2",
  "python-jose[cryptography]",
  "alembic",
  "pydantic-settings",
  "aiofiles",
  "azure-cognitiveservices-speech",
  "litellm[proxy]>=1.73.6",
  "azure-storage-blob",
  "pydub",
  "httpx",
  "arq[redis]",
  "asyncpg",
  "python-docx",
  "pypandoc",
  "ruff",
  "slowapi"
]